refactor: extract camera-ready config loader/validator into camera_ready/_config (#3385)
Continue the no-behavior-change decomposition of camera_ready_campaign.py
(issue #3385, maintainer plan Phase 3: config parsing/validation). Move the two
remaining noqa-suppressed functions out of the compatibility facade into the
canonical config owner:

- load_campaign_config (~330 LOC)
- _validate_campaign_config (~140 LOC)

Both are moved verbatim into robot_sf/benchmark/camera_ready/_config.py (the
existing config-helpers owner) and re-exported from camera_ready_campaign.py so
the legacy import/monkeypatch surface is unchanged. _resolve_path and
_resolve_observation_noise are imported lazily inside load_campaign_config to
break the _config <-> _run_state import cycle.

camera_ready_campaign.py drops from 751 to 261 LOC; the facade now holds only
thin delegating wrappers (run_campaign, prepare_campaign_preflight,
_build_route_clearance_warnings) plus re-exports and __all__.

Adds test_camera_ready_campaign_reexports_package_config_loader asserting the
re-export identity, matching the existing artifact/run-state contract tests.

No behavior change: campaign semantics, error messages, path-resolution, and
config-hash behavior are preserved (verbatim move; only a cycle-breaking lazy
import added).
